Move windows into sub directory
