+++ src/healthapp/app.py
@@ -15,16 +15,13 @@ from healthapp.user import User
 # class for start menu of the app
 class HealthApp(toga.App):
 
-    def startup(self, main_window=None):
+    def startup(self):
         # Create the apps data folder
         # self.paths.data is the directory where we can read/write to and have it persist after reboot/update etc etc.
         self.paths.data.mkdir(parents=True, exist_ok=True)
 
         # Create the main window
-        if main_window is None:
-            self.main_window = toga.MainWindow(title=self.formal_name)
-        else:
-            self.main_window = main_window
+        self.main_window = toga.MainWindow(title=self.formal_name)
 
         # Login/Register
         from healthapp.login import showAuthPage
@@ -33,12 +30,12 @@ class HealthApp(toga.App):
     def login_handler(self, user: User):
         self.user = user # Holds all user info. (see User class for details)
         print("User logged in: " + str(user))
-        self.show_main()
+        self.show_menu()
 
-    def show_main(self):
+    def show_menu(self):
         # pass self as the app instance to the ChoiceMenu class
-        from healthapp.choice_menu import ChoiceMenu
-        ChoiceMenu(self, self.app)
+        from healthapp.windows.choice_menu import ChoiceMenu
+        ChoiceMenu(self)
 
     def update_content(self, content) -> None:
         self.main_window.content = content
