+++ src/healthapp/windows/personal_details.py
@@ -8,13 +8,11 @@ from healthapp.app import HealthApp
 #-------------------------------------------------------------------------------------------------------#
 
 class PersonalDetails():
-    def __init__(self, main_window, app: HealthApp):  # accept a main_window argument
-        self.main_window = main_window 
+    def __init__(self, app: HealthApp):
         self.app = app
         self.app.update_content(self.get_content())
 
     def get_content(self) -> toga.Box:
-
         content = toga.Box(style=Pack(direction=COLUMN, background_color="#e0965e"))
         # Main box for the initial content
         header_box = toga.Box(style=Pack(direction=COLUMN, padding=(20, 20, 0)))
@@ -87,6 +85,4 @@ class PersonalDetails():
 
     def back_handler(self, widget):
         print("Back button pressed!")
-        # pass self as the app instance to the ChoiceMenu class
-        from healthapp.choice_menu import ChoiceMenu
-        ChoiceMenu(self.main_window, self.app)
+        self.app.show_menu()
